internal/arenaskl: Overflow fixes
This change fixes some reproducible instances of uint32s
overflowing (eg. in batch.memTableSize) and resulting in
undesirable behaviour, such as large batches not using
the flushableBatch code path and going into the memtable
instead. Also adds a directed panic in case we write keys/values
longer than MaxUint32 combined.

Fixes cockroachdb/cockroach#78391.

+++ mem_table.go
@@ -21,7 +21,7 @@ import (
 )
 
 func memTableEntrySize(keyBytes, valueBytes int) uint64 {
-	return uint64(arenaskl.MaxNodeSize(uint32(keyBytes)+8, uint32(valueBytes)))
+	return arenaskl.MaxNodeSize(uint32(keyBytes)+8, uint32(valueBytes))
 }
 
 // memTableEmptySize is the amount of allocated space in the arena when the
